frontend: Extract HeadlampButton from Sidebar
