+++ frontend/src/components/Sidebar/Sidebar.tsx
@@ -1,10 +1,8 @@
 import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
 import Drawer from '@material-ui/core/Drawer';
 import Grid from '@material-ui/core/Grid';
 import List from '@material-ui/core/List';
 import { makeStyles } from '@material-ui/core/styles';
-import SvgIcon from '@material-ui/core/SvgIcon';
 import clsx from 'clsx';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
@@ -13,9 +11,8 @@ import { useLocation } from 'react-router-dom';
 import { setSidebarSelected, setWhetherSidebarOpen } from '../../redux/actions/actions';
 import { useTypedSelector } from '../../redux/reducers/reducers';
 import { SidebarEntry } from '../../redux/reducers/ui';
-import { ReactComponent as LogoLight } from '../../resources/icon-light.svg';
-import { ReactComponent as LogoWithTextLight } from '../../resources/logo-light.svg';
 import CreateButton from '../common/Resource/CreateButton';
+import HeadlampButton from './HeadlampButton';
 import NavigationTabs from './NavigationTabs';
 import prepareRoutes from './prepareRoutes';
 import SidebarItem from './SidebarItem';
@@ -44,6 +41,9 @@ const useStyle = makeStyles(theme => ({
     }),
     overflowX: 'hidden',
     width: theme.spacing(7) + 1,
+    [theme.breakpoints.down('sm')]: {
+      width: theme.spacing(0),
+    },
     [theme.breakpoints.up('sm')]: {
       width: theme.spacing(9) + 1,
     },
@@ -53,22 +53,11 @@ const useStyle = makeStyles(theme => ({
     width: drawerWidth,
     background: theme.palette.sidebarBg,
   },
-  toolbar: {
-    borderBottom: '1px solid #1e1e1e',
-    paddingTop: theme.spacing(1.5),
-    paddingLeft: (props: { isSidebarOpen: boolean }) =>
-      props.isSidebarOpen ? theme.spacing(2) : theme.spacing(1),
-    paddingBottom: theme.spacing(1),
-  },
   sidebarGrid: {
     height: '100%',
   },
   '.MuiListItemText-primary': {
     color: 'red',
-  },
-  logo: {
-    height: '32px',
-    width: 'auto',
   },
 }));
 
@@ -134,8 +123,7 @@ export function PureSidebar({
   search,
   linkArea,
 }: PureSidebarProps) {
-  const classes = useStyle({ isSidebarOpen: open });
-  const { t } = useTranslation('sidebar');
+  const classes = useStyle();
 
   return (
     <Drawer
@@ -154,18 +142,7 @@ export function PureSidebar({
         component: 'nav',
       }}
     >
-      <div className={classes.toolbar}>
-        <Button
-          onClick={onToggleOpen}
-          aria-label={open ? t('Shrink sidebar') : t('Expand sidebar')}
-        >
-          <SvgIcon
-            className={classes.logo}
-            component={open ? LogoWithTextLight : LogoLight}
-            viewBox="0 0 auto 32"
-          />
-        </Button>
-      </div>
+      <HeadlampButton open={open} onToggleOpen={onToggleOpen} />
       <Grid
         className={classes.sidebarGrid}
         container
